refactor: rename and improve Erlang converter tests
