+++ tests/a2mochi/x/erl/avg_builtin.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 16:52:22 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:44 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -13,6 +13,6 @@ package main
 
 // line 6-7 (exported)
 fun main() {
-  print((sum([1, 2, 3]) / len([1, 2, 3])))
+  print(sum([1, 2, 3]) / len([1, 2, 3]))
 }
 main()
