+++ tests/a2mochi/x/erl/basic_compare.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 16:52:22 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:45 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -17,10 +17,10 @@ package main
 
 // line 6-11 (exported)
 fun main() {
-  let A = (10 - 3)
-  let B = (2 + 2)
+  let A = 10 - 3
+  let B = 2 + 2
   print(A)
-  print((A == 7))
-  print((B < 5))
+  print(A == 7)
+  print(B < 5)
 }
 main()
