+++ tests/a2mochi/x/erl/binary_precedence.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 16:52:22 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:45 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -16,9 +16,9 @@ package main
 
 // line 6-10 (exported)
 fun main() {
-  print((1 + (2 * 3)))
-  print(((1 + 2) * 3))
-  print(((2 * 3) + 1))
-  print((2 * (3 + 1)))
+  print(1 + 2 * 3)
+  print((1 + 2) * 3)
+  print(2 * 3 + 1)
+  print(2 * (3 + 1))
 }
 main()
