+++ tests/a2mochi/x/erl/map_index.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 16:52:22 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:48 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
