+++ tests/a2mochi/x/erl/string_index.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 18:17:34 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:51 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
