+++ tests/a2mochi/x/erl/string_prefix_slice.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 18:17:34 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:10 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -19,8 +19,10 @@ package main
 fun main() {
   let Prefix = "fore"
   let S1 = "forest"
-  print((substring(S1, 0, 0 + (len(Prefix) - 0)) == Prefix))
+  print("~p~n"
+  [substring(S1, 0, 0 + len(Prefix) - 0) == Prefix])
   let S2 = "desert"
-  print((substring(S2, 0, 0 + (len(Prefix) - 0)) == Prefix))
+  print("~p~n"
+  [substring(S2, 0, 0 + len(Prefix) - 0) == Prefix])
 }
 main()
