+++ tests/a2mochi/x/erl/substring_builtin.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 18:17:34 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:52 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -13,6 +13,6 @@ package main
 
 // line 6-7 (exported)
 fun main() {
-  print(substring("mochi", 1, 1 + (4 - 1)))
+  print(substring("mochi", 1, 1 + 4 - 1))
 }
 main()
