+++ tests/a2mochi/x/erl/unary_neg.mochi
@@ -1,4 +1,4 @@
-// Generated by Mochi v0.10.47 on 2025-07-28 19:28:41 GMT+7
+// Generated by Mochi v0.10.47 on 2025-07-28 22:10:53 GMT+7
 /*
 #!/usr/bin/env escript
 -module(main).
@@ -15,6 +15,6 @@ package main
 // line 6-8 (exported)
 fun main() {
   print(-3)
-  print((5 + (-2)))
+  print(5 + (-2))
 }
 main()
